feat: support fractional share quantities

+++ bots/base.py
@@ -8,7 +8,7 @@ class BaseBot(ABC):
     def __init__(self, balance: float = 1000.00, base_api_url: str = "http://127.0.0.1:8000", interval: int = 1):
         self.bot_id = str(uuid.uuid4())[:8]
         self.balance = balance
-        self.portfolio: dict[str, int] = {}
+        self.portfolio: dict[str, float] = {}
         self.interval = interval
         self.pending_orders: list[dict] = []
