+++ db/schema.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS trades(
     id UUID PRIMARY KEY,
     symbol TEXT,
     price NUMERIC,
-    quantity INTEGER,
+    quantity NUMERIC,
     buyer_order_id UUID NOT NULL,
     seller_order_id UUID NOT NULL,
     created_at TIMESTAMP DEFAULT NOW()
